database complete (#39)
* MainActivity.java fix

* Created: Bus.java
Created: User.java

* ORM database control and main classes and controllers

* fixed the database problem

* more classes for the data layer

* major changes to the file structure please be cautious when merging.

database scope and model logic extended.(incomplete)

rebase merge is recommended,

review the changes carefully..

* login fixed with hashing
new classes:
Schedule
Payment
Feedback

changes:
AppDatabase
UserDao
RegisterActivity
LoginActivity
BusDriver
BusOwner
RegisterOwnerActivity

* fixed gradle issue with new changes

* fixed gradle issue with new changes

* database is complete

+++ app/src/main/java/com/example/mad_project/controller/TicketController.java
@@ -20,12 +20,12 @@ public class TicketController {
 
     public void getTicketPriceByBusId(int busId, TicketPriceCallback callback) {
         executorService.execute(() -> {
-            int ticketPrice = ticketDao.getTicketPriceByBusId(busId);
+            double ticketPrice = ticketDao.getTicketPriceByBusId(busId);
             callback.onTicketPriceLoaded(ticketPrice);
         });
     }
 
     public interface TicketPriceCallback {
-        void onTicketPriceLoaded(int ticketPrice);
+        void onTicketPriceLoaded(double ticketPrice);
     }
 }
